Few tweaks on dashboard

+++ src/pages/Dashboard/ManageProduct/ManageProduct.js
@@ -29,6 +29,9 @@ const ManageProduct = () => {
   };
 
   const handleDelete = async (deleteId) => {
+    if (!window.confirm("Are you sure you want to delete?")) {
+      return;
+    }
     const toastId = toast.loading("Deleting product");
     const result = await deleteSingleProduct(deleteId);
